[ci skip] [WIP] Fix build/test issues for entire codebase

+++ src/jet/collocated_vector_grid2.cpp
@@ -111,7 +111,8 @@ VectorGrid2::ConstVectorDataView CollocatedVectorGrid2::dataView() const {
 VectorGrid2::DataPositionFunc CollocatedVectorGrid2::dataPosition() const {
     Vector2D dataOrigin_ = dataOrigin();
     return [this, dataOrigin_](size_t i, size_t j) -> Vector2D {
-        return dataOrigin_ + elemMul(gridSpacing(), Vector2D(i, j));
+        return dataOrigin_ +
+               elemMul(gridSpacing(), Vector2D((double)i, (double)j));
     };
 }
